feat: Renamed pages
